requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.6.2
channels==4.0.0
click==8.1.3
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distro==1.8.0
Django==4.2.2
django-bootstrap4==23.1
django-multiupload==0.6.1
django-plotly-dash==2.2.1
dpd-components==0.1.0
dpd-static-support==0.0.5
et-xmlfile==1.1.0
fire==0.5.0
Flask==2.3.2
fonttools==4.40.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.3
numpy==1.25.0
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pdf2docx==0.5.6
Pillow==10.0.0
plotly==5.15.0
PyMuPDF==1.22.5
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tabula-py==2.7.0
tenacity==8.2.2
termcolor==2.3.0
typing_extensions==4.6.3
tzdata==2023.3
Werkzeug==2.3.6
whitenoise==6.5.0
